requirements.txt
streamlit
transformers
torch
accelerate
bitsandbytes
peft
python-dotenv